requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
pip-tools==6.9.0
    # via -r requirements.in
kafka-python==2.0.2
    # via -r requirements.in
dnspython==2.2.1 
pymongo==4.3.2